Parallelize ForAllSnapshots
